Second phase of migration
- Remove all code related to the backend queries
- Rename and clarify what is left for the front
  end rewrite
